Fix editProduct function
A new event listener was added to the update product button
every time the user clicked the button and triggered the
editProduct function. This caused an accumulation of event
listeners on the button which would trigger the updateProduct
function repeatedly. To fix this, the "{ once: true }" parameter
was added to the addEventListener method call.